requirements.txt
amqp==5.0.6
asgiref==3.4.1
astroid==2.7.2
billiard==3.6.4.0
celery==5.1.2
certifi==2021.5.30
charset-normalizer==2.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
Django==3.2.6
django-extensions==3.1.3
Faker==8.11.0
flake8==3.9.2
flake8-builtins==1.5.3
flake8-import-order==0.18.1
flake8-print==4.0.0
idna==3.2
isort==5.9.3
kombu==5.1.0
lazy-object-proxy==1.6.0
mccabe==0.6.1
platformdirs==2.2.0
prompt-toolkit==3.0.20
pycodestyle==2.7.0
pyflakes==2.3.1
pylint==2.10.2
pylint-django==2.4.4
pylint-plugin-utils==0.6
pytest-django==4.4.0
python-dateutil==2.8.2
pytz==2021.1
requests==2.26.0
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
urllib3==1.26.6
vine==5.0.0
wcwidth==0.2.5
wrapt==1.12.1
